requirements.txt
requests
websocket-client
loguru